+++ src/main/java/net/silentchaos512/gems/chaos/ChaosEvents.java
@@ -3,51 +3,50 @@ package net.silentchaos512.gems.chaos;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.effect.EntityLightningBolt;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraft.world.gen.Heightmap;
 import net.silentchaos512.gems.SilentGems;
+import net.silentchaos512.gems.api.chaos.ChaosEvent;
 import net.silentchaos512.gems.block.CorruptedBlocks;
 import net.silentchaos512.utils.MathUtils;
 import org.apache.logging.log4j.Marker;
 import org.apache.logging.log4j.MarkerManager;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.function.BiConsumer;
+import java.util.HashMap;
+import java.util.Map;
 
 public final class ChaosEvents {
-    private static final Collection<BiConsumer<Entity, Integer>> EVENT_LIST = new ArrayList<>();
+    private static final Map<ResourceLocation, ChaosEvent> EVENTS = new HashMap<>();
     private static final Marker MARKER = MarkerManager.getMarker("ChaosEvents");
 
     // Not an actual cap on the value, just used for event probability
-    private static final int MAX_CHAOS = 10_000_000;
+    private static final int MAX_CHAOS = 5_000_000;
 
     static {
-        EVENT_LIST.add((entity, chaos) -> {
-            if (chaos > 100_000 && tryChance(0.25f, chaos, MAX_CHAOS)) {
-                SilentGems.LOGGER.info(MARKER, "Chaos lightning @ {}", entity);
-                spawnLightningBolt(entity, entity.world);
-                Chaos.dissipate(entity.world, 10_000);
-            }
-        });
-        EVENT_LIST.add((entity, chaos) -> {
-            if (chaos > 1_000_000 && tryChance(0.25f, chaos, MAX_CHAOS)) {
-                SilentGems.LOGGER.info(MARKER, "Creating corrupted blocks @ {}", entity);
-                corruptBlocks(entity, entity.world);
-                Chaos.dissipate(entity.world, 100_000);
-            }
-        });
+        addChaosEvent(SilentGems.getId("chaos_lightning"), new ChaosEvent(
+                0.25f, 150_000, MAX_CHAOS, 10_000, (entity, chaos) ->
+                spawnLightningBolt(entity, entity.world)
+        ));
+        addChaosEvent(SilentGems.getId("corrupt_blocks"), new ChaosEvent(
+                0.25f, 1_000_000, MAX_CHAOS, 100_000, (entity, chaos) ->
+                corruptBlocks(entity, entity.world)
+        ));
     }
 
     private ChaosEvents() {throw new IllegalAccessError("Utility class");}
 
-    public static void addChaosEvent(BiConsumer<Entity, Integer> action) {
-        EVENT_LIST.add(action);
+    public static void addChaosEvent(ResourceLocation id, ChaosEvent event) {
+        EVENTS.put(id, event);
     }
 
     static void tryChaosEvents(Entity entity, World world, int chaos) {
-        EVENT_LIST.forEach(consumer -> consumer.accept(entity, chaos));
+        EVENTS.forEach((id, event) -> {
+            if (event.tryActivate(entity, chaos)) {
+                SilentGems.LOGGER.info(MARKER, "Activate {} @ {}", id, entity.getScoreboardName());
+            }
+        });
     }
 
     private static boolean tryChance(float max, int chaos, int maxChaos) {