+++ src/main/java/net/silentchaos512/gems/client/ClientPlayerInfo.java
@@ -3,6 +3,7 @@ package net.silentchaos512.gems.client;
 public final class ClientPlayerInfo {
     public static int playerChaos;
     public static int worldChaos;
+    public static int equilibriumChaos;
 
     private ClientPlayerInfo() {throw new IllegalAccessError("Utility class");}
 }