+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- World chaos will now balance at a certain level, instead of zero. The level fluctuates over time. 
 ### Fixed
 - Crash when Silent Gear not installed [#365]